+++ redflag/independence.py
@@ -1,0 +1,67 @@
+"""
+Functions related to understanding row independence.
+
+Author: Matt Hall, agilescientific.com
+Licence: Apache 2.0
+
+Copyright 2022 Agile Scientific
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import numpy as np
+
+
+def is_correlated(a, n=20, s=20, threshold=0.1):
+    """
+    Check if a dataset is correlated. Uses s chunks of n samples.
+
+    Args:
+        a (array): The data.
+        n (int): The number of samples per chunk.
+        s (int): The number of chunks.
+        threshold (float): The auto-correlation threshold.
+
+    Returns:
+        bool: True if the data are correlated.
+
+    Examples:
+        >>> is_correlated([7, 1, 6, 8, 7, 6, 2, 9, 4, 2])
+        False
+        >>> is_correlated([1, 2, 1, 7, 6, 8, 6, 2, 1, 1])
+        True
+    """
+    a = np.asarray(a)
+
+    # Split into chunks n samples long.
+    L_chunks = min(a.size, n)
+    chunks = np.array_split(a, a.size//L_chunks)
+
+    # Choose up to s chunk indices at random.
+    N_chunks = min(len(chunks), s)
+    rng = np.random.default_rng()
+    r = rng.choice(np.arange(len(chunks)), size=N_chunks, replace=False)
+
+    # Loop over selected chunks and count ones with correlation.
+    acs = []
+    for chunk in [c for i, c in enumerate(chunks) if i in r]:
+        c = chunk[:L_chunks] - np.nanmean(chunk)
+        autocorr = np.correlate(c, c, mode='same')
+        acs.append(autocorr / (c.size * np.nanvar(c)))
+
+    # Average the autocorrelations.
+    acs = np.sum(acs, axis=0) / N_chunks
+
+    p = acs[c.size//2 - 1]  # First non-zero lag.
+    q = acs[c.size//2 - 2]  # Next non-zero lag.
+
+    return (p >= threshold) & (q >= 0)